+++ esp-hal-smartled/examples/hello_rgb.rs
@@ -84,6 +84,7 @@ fn main() -> ! {
             rmt.channel0,
             led_pin,
             2,
+            freq
         ).unwrap()
     };
     let delay = Delay::new();